+++ src/refactoring/hooks/admin/index.ts
@@ -1,0 +1,1 @@
+export * from "./useProductForm";
